# Add gunzip, gzip to lagom and make deflate decompression a bit faster on arm