@nico 2023-03-25
(Slightly less flippantly: This reduces time in crc32 from 13% of runtime to 0.6% of runtime, so it _is_ a somewhat decent speedup of a smaller part of the computation.)